Export the PARAFAC model

Currently only supports 3-mode models.

Usage

exportPARAFAC(
  model,
  dataset,
  prefix = "PARAFACmodel_",
  path = "./",
  saveRDS = FALSE
)

Arguments

model

Model output from parafac().

dataset

A longitudinal microbiome dataset, ideally processed with processDataCube(), formatted as follows:

data

Array object of the data cube

mode1

Dataframe with all the subject metadata, ordered the same as the rows in the data cube.

mode2

Taxonomic classification of the microbiota, ordered the same as the columns in the data cube.

mode3

Dataframe with the time metadata, ordered the same as the third dimension in the array.

prefix

Prefix file name

path

Path to a folder

saveRDS

Save an RDS of the PARAFAC model object? (true/false, default=FALSE)

Value

Saves mode1, mode2, mode3, input data, modelled data, and data as modelled per component in .csv files.

Examples

library(multiway)
library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
library(ggplot2)
library(paramGUI)
library(pracma)
set.seed(0)

# Make PARAFAC model
processedFujita = processDataCube(Fujita2023, sparsityThreshold=0.99, centerMode=1, scaleMode=2)
model = parafac(processedFujita$data, nfac=2, nstart=1, verbose=FALSE)
if (FALSE) {
exportPARAFAC(model, processedFujita, prefix="Fujita_")
}