parafac4microbiome: Parallel Factor Analysis Modelling of Longitudinal Microbiome Data

Creation and selection of PARAllel FACtor Analysis (PARAFAC) models of longitudinal microbiome data. You can import your own data with our import functions or use one of the example datasets to create your own PARAFAC models. Selection of the optimal number of components can be done using assessModelQuality() and assessModelStability(). The selected model can then be plotted using plotPARAFACmodel(). The Parallel Factor Analysis method was originally described by Caroll and Chang (1970) doi:10.1007/BF02310791 and Harshman (1970) https://www.psychology.uwo.ca/faculty/harshman/wpppfac0.pdf.

See also

Author

Maintainer: Geert Roelof van der Ploeg g.r.ploeg@uva.nl (ORCID)

Other contributors:

Johan Westerhuis j.a.westerhuis@uva.nl (ORCID) [contributor]

Anna Heintz-Buschart a.u.s.heintzbuschart@uva.nl (ORCID) [contributor]

Age Smilde a.k.smilde@uva.nl (ORCID) [contributor]

University of Amsterdam [copyright holder, funder]