vanderPloeg2024 longitudinal microbiome data

The vanderPloeg2024 longitudinal microbiome dataset as a three-dimensional array, with subjects in mode 1, microbial abundances in mode 2, and time in mode 3. Note: all-zero microbial abundances have been removed to save disk space.

Usage

vanderPloeg2024

Format

vanderPloeg2024

A list object with three elements:

data

Array object of the data cube

mode1

Dataframe with all the subject metadata, ordered the same as the rows in the data cube.

mode2

Taxonomic classification of the microbiota, ordered the same as the columns in the data cube.

mode3

Dataframe with the time metadata, ordered the same as the third dimension in the array.

...

Source

doi:10.1101/2024.03.18.585469