Parallel Factor Analysis

Parallel Factor Analysis

Usage

parafac(
  Tensor,
  nfac,
  nstart = 1,
  maxit = 500,
  ctol = 1e-04,
  initialization = "random",
  verbose = FALSE,
  output = "best"
)

Arguments

Tensor

3-way matrix of numeric data

nfac

Number of factors (components) to fit.

nstart

Number of models to randomly initialize (default 1).

maxit

Maximum number of iterations allowed without convergence (default 500).

ctol

Relative change in loss tolerated to call the algorithm converged (default 1e-4).

initialization

"Random" for randomly initialized input vectors or "nvec" for svd-based best guess.

verbose

DEPRECATED: verbose output

output

String ("best"/"all") Return only the best model of the nstart models ("best") or return all of them in a list object ("all").

Value

List object of the PARAFAC model or models.

Examples

X = array(rnorm(108*100*10), c(108,100,10))
model = parafac(X, 2)
#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%