Parallel Factor Analysis-1

Fits Richard A. Harshman's Parallel Factors (Parafac) model to 3-way or 4-way data arrays. Parameters are estimated via alternating least squares with optional constraints.

Usage

parafac(
  X,
  nfac,
  nstart = 10,
  const = NULL,
  control = NULL,
  Afixed = NULL,
  Bfixed = NULL,
  Cfixed = NULL,
  Dfixed = NULL,
  Astart = NULL,
  Bstart = NULL,
  Cstart = NULL,
  Dstart = NULL,
  Astruc = NULL,
  Bstruc = NULL,
  Cstruc = NULL,
  Dstruc = NULL,
  Amodes = NULL,
  Bmodes = NULL,
  Cmodes = NULL,
  Dmodes = NULL,
  maxit = 500,
  ctol = 1e-04,
  parallel = FALSE,
  cl = NULL,
  output = c("best", "all"),
  verbose = TRUE,
  backfit = FALSE
)

Arguments

X

Three-way data array with dim=c(I,J,K) or four-way data array with dim=c(I,J,K,L). Missing data are allowed (see Note).

nfac

Number of factors.

nstart

Number of random starts.

const

Character vector of length 3 or 4 giving the constraints for each mode (defaults to unconstrained). See CMLS::const for the 24 available options.

control

List of parameters controlling options for smoothness constraints. This is passed to multiway::const.control, which describes the available options.

Afixed

Used to fit model with fixed Mode A weights.

Bfixed

Used to fit model with fixed Mode B weights.

Cfixed

Used to fit model with fixed Mode C weights.

Dfixed

Used to fit model with fixed Mode D weights.

Astart

Starting Mode A weights. Default uses random weights.

Bstart

Starting Mode B weights. Default uses random weights.

Cstart

Starting Mode C weights. Default uses random weights.

Dstart

Starting Mode D weights. Default uses random weights.

Astruc

Structure constraints for Mode A weights. See Note.

Bstruc

Structure constraints for Mode B weights. See Note.

Cstruc

Structure constraints for Mode C weights. See Note.

Dstruc

Structure constraints for Mode D weights. See Note.

Amodes

Mode ranges for Mode A weights (for unimodality constraints). See Note.

Bmodes

Mode ranges for Mode B weights (for unimodality constraints). See Note.

Cmodes

Mode ranges for Mode C weights (for unimodality constraints). See Note.

Dmodes

Mode ranges for Mode D weights (for unimodality constraints). See Note.

maxit

Maximum number of iterations.

ctol

Convergence tolerance (R^2 change).

parallel

Logical indicating if parallel::parLapply should be used. See Examples.

cl

Cluster created by parallel::makeCluster. Only used when parallel=TRUE.

output

Output the best solution (default) or output all nstart solutions.

verbose

If TRUE, fitting progress is printed via utils::txtProgressBar. Ignored if parallel=TRUE.

backfit

Should backfitting algorithm be used for CMLS::cmls?

Value

If output = "best", returns an object of class "parafac" with the following elements: #'

A

Mode A weight matrix.

B

Mode B weight matrix.

C

Mode C weight matrix.

D

Mode D weight matrix.

SSE

Sum of Squared Errors.

Rsq

R-squared value.

GCV

Generalized Cross-Validation.

edf

Effective degrees of freedom.

iter

Number of iterations.

cflag

Convergence flag. See Note.

const

See argument const.

control

See argument control.

fixed

Logical vector indicating whether 'fixed' weights were used for each mode.

struc

Logical vector indicating whether 'struc' constraints were used for each mode.

Otherwise returns a list of length nstart where each element is an object of class "parafac".

Note

Missing data should be specified as NA values in the input X. The missing data are randomly initialized and then iteratively imputed as a part of the algorithm.

Structure constraints should be specified with a matrix of logicals (TRUE/FALSE), such that FALSE elements indicate a weight should be constrained to be zero. Default uses unstructured weights, i.e., a matrix of all TRUE values.

When using unimodal constraints, the *modes inputs can be used to specify the mode search range for each factor. These inputs should be matrices with dimension c(2,nfac) where the first row gives the minimum mode value and the second row gives the maximum mode value (with respect to the indicies of the corresponding weight matrix).

Output cflag gives convergence information: cflag = 0 if algorithm converged normally, cflag = 1 if maximum iteration limit was reached before convergence, and cflag = 2 if algorithm terminated abnormally due to a problem with the constraints.

Examples

library(multiway)
#' ##########   3-way example   ##########

# create random data array with Parafac structure
set.seed(3)
mydim <- c(50, 20, 5)
nf <- 3
Amat <- matrix(rnorm(mydim[1]*nf), nrow = mydim[1], ncol = nf)
Bmat <- matrix(runif(mydim[2]*nf), nrow = mydim[2], ncol = nf)
Cmat <- matrix(runif(mydim[3]*nf), nrow = mydim[3], ncol = nf)
Xmat <- tcrossprod(Amat, krprod(Cmat, Bmat))
Xmat <- array(Xmat, dim = mydim)
Emat <- array(rnorm(prod(mydim)), dim = mydim)
Emat <- nscale(Emat, 0, ssnew = sumsq(Xmat))   # SNR = 1
X <- Xmat + Emat

# fit Parafac model (unconstrained)
pfac <- parafac(X, nfac = nf, nstart = 1)
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
pfac
#> 
#> 3-way Parafac with 3 factors 
#> 
#> Constraints:
#>       A      B      C
#>  uncons uncons uncons
#> 
#> Fit Information:
#>   SSE = 950.3324 
#>   R^2 = 0.5149861 
#>   GCV = 0.2078778 
#>   EDF = 219 
#>  
#> Converged: TRUE (15 iterations)
#>  

# fit Parafac model (non-negativity on Modes B and C)
pfacNN <- parafac(X, nfac = nf, nstart = 1,
            const = c("uncons", "nonneg", "nonneg"))
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
pfacNN
#> 
#> 3-way Parafac with 3 factors 
#> 
#> Constraints:
#>       A      B      C
#>  uncons nonneg nonneg
#> 
#> Fit Information:
#>   SSE = 950.1934 
#>   R^2 = 0.515057 
#>   GCV = 0.2078474 
#>   EDF = 219 
#>  
#> Converged: TRUE (12 iterations)
#>  

# check solution
Xhat <- fitted(pfac)
sum((Xmat - Xhat)^2) / prod(mydim)
#> [1] 0.008264583

# reorder and resign factors
pfac$B[1:4,]
#>            [,1]       [,2]       [,3]
#> [1,] -0.2993941 -0.9361133 -0.3507422
#> [2,] -0.3731378 -0.3491355 -0.4316282
#> [3,] -1.4244651 -0.8101654 -1.0452068
#> [4,] -0.9485296 -0.8058645 -0.5985801
pfac <- reorder(pfac, c(3,1,2))
pfac$B[1:4,]
#>            [,1]       [,2]       [,3]
#> [1,] -0.3507422 -0.2993941 -0.9361133
#> [2,] -0.4316282 -0.3731378 -0.3491355
#> [3,] -1.0452068 -1.4244651 -0.8101654
#> [4,] -0.5985801 -0.9485296 -0.8058645
pfac <- resign(pfac, mode="B")
pfac$B[1:4,]
#>           [,1]      [,2]      [,3]
#> [1,] 0.3507422 0.2993941 0.9361133
#> [2,] 0.4316282 0.3731378 0.3491355
#> [3,] 1.0452068 1.4244651 0.8101654
#> [4,] 0.5985801 0.9485296 0.8058645
Xhat <- fitted(pfac)
sum((Xmat - Xhat)^2) / prod(mydim)
#> [1] 0.008264583

# rescale factors
colSums(pfac$B^2)
#> [1] 20 20 20
colSums(pfac$C^2)
#> [1] 0.2374967 0.6032392 0.2727013
pfac <- rescale(pfac, mode = "C", absorb = "B")
colSums(pfac$B^2)
#> [1] 0.9499866 2.4129567 1.0908051
colSums(pfac$C^2)
#> [1] 5 5 5
Xhat <- fitted(pfac)
sum((Xmat - Xhat)^2) / prod(mydim)
#> [1] 0.008264583

##########   4-way example   ##########

# create random data array with Parafac structure
set.seed(4)
mydim <- c(30,10,8,10)
nf <- 4
aseq <- seq(-3, 3, length.out = mydim[1])
Amat <- cbind(dnorm(aseq), dchisq(aseq+3.1, df=3),
      dt(aseq-2, df=4), dgamma(aseq+3.1, shape=3, rate=1))
Bmat <- svd(matrix(runif(mydim[2]*nf), nrow = mydim[2], ncol = nf), nv = 0)$u
Cmat <- matrix(runif(mydim[3]*nf), nrow = mydim[3], ncol = nf)
Cstruc <- Cmat > 0.5
Cmat <- Cmat * Cstruc
Dmat <- matrix(runif(mydim[4]*nf), nrow = mydim[4], ncol = nf)
Xmat <- tcrossprod(Amat, krprod(Dmat, krprod(Cmat, Bmat)))
Xmat <- array(Xmat, dim = mydim)
Emat <- array(rnorm(prod(mydim)), dim = mydim)
Emat <- nscale(Emat, 0, ssnew = sumsq(Xmat))   # SNR = 1
X <- Xmat + Emat

# fit Parafac model (unimodal and smooth A, orthogonal B,
#                    non-negative and structured C, non-negative D)
pfac <- parafac(X, nfac = nf, nstart = 1, Cstruc = Cstruc,
            const = c("unismo", "orthog", "nonneg", "nonneg"))
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
pfac
#> 
#> 4-way Parafac with 4 factors 
#> 
#> Constraints:
#>       A      B             C      D
#>  unismo orthog nonneg+struct nonneg
#> 
#> Fit Information:
#>   SSE = 42.48407 
#>   R^2 = 0.4873829 
#>   GCV = 0.001785761 
#>   EDF = 105 
#>  
#> Converged: TRUE (7 iterations)
#>  

# check solution
Xhat <- fitted(pfac)
sum((Xmat - Xhat)^2) / prod(mydim)
#> [1] 5.132859e-05
congru(Amat, pfac$A)
#>           [,1]      [,2]      [,3]       [,4]
#> [1,] 0.9996716 0.7287234 0.4135958 -0.8044008
#> [2,] 0.7123512 0.9990610 0.3778488 -0.9789737
#> [3,] 0.4145023 0.3786920 0.9986611 -0.4571581
#> [4,] 0.8712777 0.9374104 0.4200941 -0.9332070
crossprod(pfac$B)
#>               [,1]         [,2]          [,3]          [,4]
#> [1,]  1.000000e+01 3.345089e-15  9.014050e-16 -1.413387e-15
#> [2,]  3.345089e-15 1.000000e+01  1.207936e-15  2.749057e-15
#> [3,]  9.014050e-16 1.207936e-15  1.000000e+01 -4.394494e-16
#> [4,] -1.413387e-15 2.749057e-15 -4.394494e-16  1.000000e+01
pfac$C
#>              [,1]      [,2]          [,3]      [,4]
#> [1,] 1.429197e+00 1.0865265 -1.062753e-19 0.0000000
#> [2,] 7.471576e-19 0.8404925  6.669612e-36 1.2003143
#> [3,] 8.760174e-01 0.0000000  1.477234e+00 0.0000000
#> [4,] 0.000000e+00 1.1272930  1.090825e+00 0.8922219
#> [5,] 1.342355e+00 1.3602890  9.173661e-01 1.3893746
#> [6,] 1.709773e-19 1.1615836  0.000000e+00 1.6053098
#> [7,] 1.289349e+00 1.2816319  1.307693e+00 0.0000000
#> [8,] 1.313641e+00 0.0000000  1.440923e+00 1.1206267
Cstruc
#>       [,1]  [,2]  [,3]  [,4]
#> [1,]  TRUE  TRUE FALSE FALSE
#> [2,] FALSE  TRUE FALSE  TRUE
#> [3,]  TRUE FALSE  TRUE FALSE
#> [4,] FALSE  TRUE  TRUE  TRUE
#> [5,]  TRUE  TRUE  TRUE  TRUE
#> [6,] FALSE  TRUE FALSE  TRUE
#> [7,]  TRUE  TRUE  TRUE FALSE
#> [8,]  TRUE FALSE  TRUE  TRUE